+++ sdk/attestation/Azure.Security.Attestation/src/Models/PolicyCertificatesModificationResult.cs
@@ -5,10 +5,11 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Text.Json.Serialization;
+using Azure.Core;
 
 namespace Azure.Security.Attestation.Models
 {
-    [JsonConverter(typeof(PolicyCertificateModificationResultConverter))]
+    [CodeGenModel(Usage = new[] { "converter" })]
     public partial class PolicyCertificatesModificationResult
     {
     }