+++ sdk/attestation/Azure.Security.Attestation/tests/Samples/AttestationServiceAttestationSamples.cs
@@ -141,7 +141,7 @@ namespace Azure.Security.Attestation.Tests.Samples
             IReadOnlyList<AttestationSigner> signingCertificates = attestClient.GetSigningCertificates().Value;
 
             var policyResult = await client.GetPolicyAsync(AttestationType.SgxEnclave);
-            var result = policyResult.Value.AttestationPolicy;
+            var result = policyResult.Value;
         }
 
         [Test]
@@ -152,7 +152,7 @@ namespace Azure.Security.Attestation.Tests.Samples
             var client = new AttestationAdministrationClient(new Uri(endpoint), new DefaultAzureCredential());
 
             var policyResult = await client.GetPolicyAsync(AttestationType.SgxEnclave);
-            var result = policyResult.Value.AttestationPolicy;
+            var result = policyResult.Value;
             #endregion
 
 #region Snippet:SetPolicy
@@ -160,19 +160,14 @@ namespace Azure.Security.Attestation.Tests.Samples
 
             var policyTokenSigner = TestEnvironment.PolicyCertificate0;
 
-            AttestationToken policySetToken = new SecuredAttestationToken(
-                new StoredAttestationPolicy { AttestationPolicy = attestationPolicy, },
-                TestEnvironment.PolicySigningKey0,
-                policyTokenSigner);
-
-            var setResult = client.SetPolicy(AttestationType.SgxEnclave, policySetToken);
+            var setResult = client.SetPolicy(AttestationType.SgxEnclave, attestationPolicy, TestEnvironment.PolicySigningKey0, policyTokenSigner);
 #endregion
             var resetResult = client.ResetPolicy(AttestationType.SgxEnclave);
 
             // When the attestation instance is in Isolated mode, the ResetPolicy API requires using a signing key/certificate to authorize the user.
             var resetResult2 = client.ResetPolicy(
                 AttestationType.SgxEnclave,
-                new SecuredAttestationToken(TestEnvironment.PolicySigningKey0, policyTokenSigner));
+                TestEnvironment.PolicySigningKey0, policyTokenSigner);
             return;
         }
         private AttestationClient GetAttestationClient()