+++ sdk/attestation/Azure.Security.Attestation/CHANGELOG.md
@@ -5,6 +5,10 @@
 - Fixed bug #19708, handle JSON values that are not just simple integers.
 - Fixed bug #18183, Significant cleanup of README.md.
 
+### Breaking Changes:
+It is no longer necessary to manually Base64Url encode the AttestationPolicy property in the StoredAttestationPolicy model. 
+This simplifies the user experience for interacting with the saved attestation policies - developers can treat attestation policies as string values.
+
 ## 1.0.0-beta.1 (2021-01-15)
 Released as beta, not alpha.